Code review suggestions done
